refactor: move button generation helpers to helpers specific file, replace IsAlive with reverted IsDead
